OpenMP: Refactor files into parallel patterns.
